pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta" 

[project]
name = "gum-ai"
version = "0.1.11"
readme = "README.md"
authors = [{ name = "Omar Shaikh", email = "oshaikh13@gmail.com" }]
license = {text = "MIT"}
dependencies = [
    "pillow",
    "mss",
    "pynput",
    "pyobjc-framework-Quartz; sys_platform == 'darwin'",
    "shapely; sys_platform == 'darwin'",
    "openai>=1.0.0",
    "SQLAlchemy>=2.0.0",
    "pydantic>=2.0.0",
    "sqlalchemy-utils>=0.41.0",
    "python-dotenv>=1.0.0",
    "scikit-learn",
    "aiosqlite",
    "aiohttp",
    "beautifulsoup4",
    "ics",
    "tatsu<5.0",
    "greenlet",
    "persist-queue",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings>=0.24.0",
    "mkdocstrings-python>=1.7.0"
]
requires-python = ">=3.6"

[project.scripts]
gum = "gum.cli:cli"

[project.urls]
"Homepage" = "https://github.com/GeneralUserModels/gum"